backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# ── Caching ───────────────────────────────────────────────────────────────────
redis==5.0.4
aioredis==2.0.1

# ── ML / AI ──────────────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
transformers==4.40.0
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-learn==1.4.2
scipy==1.13.0

# ── HTTP / Networking ─────────────────────────────────────────────────────────
httpx==0.27.0
websockets==12.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0

# ── Logging & Monitoring ──────────────────────────────────────────────────────
structlog==24.2.0
sentry-sdk[fastapi]==2.2.0
prometheus-fastapi-instrumentator==7.0.0

deepface
tf-keras
librosa==0.10.2
soundfile==0.12.1
